refactor: Implement modules for components and states
